+++ lib/src/main/java/com/sweetzpot/stravazpot/activity/request/ListActivityZonesRequest.java
@@ -10,11 +10,11 @@ import retrofit2.Call;
 
 public class ListActivityZonesRequest {
 
-    private final int activityID;
+    private final long activityID;
     private final ActivityRest restService;
     private final ActivityAPI api;
 
-    public ListActivityZonesRequest(int activityID, ActivityRest restService, ActivityAPI api) {
+    public ListActivityZonesRequest(long activityID, ActivityRest restService, ActivityAPI api) {
         this.activityID = activityID;
         this.restService = restService;
         this.api = api;
