+++ lib/src/main/java/com/sweetzpot/stravazpot/activity/request/ListRelatedActivitiesRequest.java
@@ -10,13 +10,13 @@ import retrofit2.Call;
 
 public class ListRelatedActivitiesRequest {
 
-    private final int activityID;
+    private final long activityID;
     private final ActivityRest restService;
     private final ActivityAPI api;
     private Integer page;
     private Integer perPage;
 
-    public ListRelatedActivitiesRequest(int activityID, ActivityRest restService, ActivityAPI api) {
+    public ListRelatedActivitiesRequest(long activityID, ActivityRest restService, ActivityAPI api) {
         this.activityID = activityID;
         this.restService = restService;
         this.api = api;
