+++ lib/src/main/java/com/sweetzpot/stravazpot/activity/rest/CommentsRest.java
@@ -12,7 +12,7 @@ import retrofit2.http.Query;
 public interface CommentsRest {
     @GET("activities/{id}/comments")
     Call<List<Comment>> getActivityComments(
-            @Path("id") Integer id,
+            @Path("id") Long id,
             @Query("page") Integer page,
             @Query("per_page") Integer perPage);
 }
