+++ lib/src/main/java/com/sweetzpot/stravazpot/athlete/request/GetBothFollowingRequest.java
@@ -10,13 +10,13 @@ import retrofit2.Call;
 
 public class GetBothFollowingRequest {
 
-    private final int athleteID;
+    private final long athleteID;
     private final FriendRest restService;
     private final FriendAPI api;
     private Integer page;
     private Integer perPage;
 
-    public GetBothFollowingRequest(int athleteID, FriendRest restService, FriendAPI api) {
+    public GetBothFollowingRequest(long athleteID, FriendRest restService, FriendAPI api) {
         this.athleteID = athleteID;
         this.restService = restService;
         this.api = api;
