+++ lib/src/main/java/com/sweetzpot/stravazpot/club/request/LeaveClubRequest.java
@@ -8,11 +8,11 @@ import retrofit2.Call;
 
 public class LeaveClubRequest {
 
-    private final int clubID;
+    private final long clubID;
     private final ClubRest restService;
     private final ClubAPI api;
 
-    public LeaveClubRequest(int clubID, ClubRest restService, ClubAPI api) {
+    public LeaveClubRequest(long clubID, ClubRest restService, ClubAPI api) {
         this.clubID = clubID;
         this.restService = restService;
         this.api = api;
