+++ lib/src/main/java/com/sweetzpot/stravazpot/club/request/ListClubAnnouncementsRequest.java
@@ -10,11 +10,11 @@ import retrofit2.Call;
 
 public class ListClubAnnouncementsRequest {
 
-    private final int clubID;
+    private final long clubID;
     private final ClubRest restService;
     private final ClubAPI api;
 
-    public ListClubAnnouncementsRequest(int clubID, ClubRest restService, ClubAPI api) {
+    public ListClubAnnouncementsRequest(long clubID, ClubRest restService, ClubAPI api) {
         this.clubID = clubID;
         this.restService = restService;
         this.api = api;
