+++ lib/src/main/java/com/sweetzpot/stravazpot/segment/request/StarSegmentRequest.java
@@ -8,12 +8,12 @@ import retrofit2.Call;
 
 public class StarSegmentRequest {
 
-    private final int segmentID;
+    private final long segmentID;
     private final boolean star;
     private final SegmentRest restService;
     private final SegmentAPI api;
 
-    public StarSegmentRequest(int segmentID, boolean star, SegmentRest restService, SegmentAPI api) {
+    public StarSegmentRequest(long segmentID, boolean star, SegmentRest restService, SegmentAPI api) {
         this.segmentID = segmentID;
         this.star = star;
         this.restService = restService;
